establishes a mock fs, and some tests for fs functions, and modularizes the router functins
