+++ package.json
@@ -7,7 +7,7 @@
     "jsdoc": "jsdoc -c ./docs/config/jsdoc.config.json",
     "lint": "eslint **/*.js",
     "test-watch": "jest --watchAll",
-    "test": "jest",
+    "test": "jest --verbose --coverage",
     "start": "node index.js"
   },
   "repository": {
@@ -33,7 +33,6 @@
     "cors": "2.8.5",
     "dotenv": "6.2.0",
     "express": "4.16.4",
-    "fs": "0.0.1-security",
     "jsdoc": "^3.5.5",
     "jsonwebtoken": "^8.4.0",
     "mongodb-memory-server": "2.9.1",
